Change the organization of nzw
## What changes were proposed in this pull request?

Update Kento Nozawa's organization.

## What type of PR is it?

[Documentation]

## What is the Jira issue?

This fix is a minor fix for release, so I have not yet created a jira ticket. Let me know if I should create it.

## How was this patch tested?

By compiling the source code, confirming this change does not break the structure of pom.xml.

## Checklist

(Please remove this section if not needed; check `x` for YES, blank for NO)

- [x] Did you apply source code formatter, i.e., `mvn formatter:format`, for your commit?
- [ ] Did you run system tests on Hive (or Spark)?

Author: Kento NOZAWA <k_nzw@klis.tsukuba.ac.jp>

Closes #140 from nzw0301/change-nzw-org.