[HIVEMALL-189] Create a list of all functions
## What changes were proposed in this pull request?

Create a list of all functions in the documentation. In order to make maintenance easier and simpler, the list is systematically generated by reading `Description` annotation in the code: [takuti/hivemalldoc](https://github.com/takuti/hivemalldoc).

In case this list does not look sufficient, let's update `Description` annotation itself and make the code more informative in the future.

## What type of PR is it?

Documentation

## What is the Jira issue?

https://issues.apache.org/jira/browse/HIVEMALL-189

Author: Takuya Kitazawa <k.takuti@gmail.com>

Closes #143 from takuti/HIVEMALL-189.